add atof, isxdigit

+++ ft_strnlen.c
@@ -6,13 +6,13 @@
 /*   By: wvenita <wvenita@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/08 23:58:50 by wvenita           #+#    #+#             */
-/*   Updated: 2019/04/09 00:18:56 by wvenita          ###   ########.fr       */
+/*   Updated: 2019/04/15 22:10:46 by wvenita          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strnlen(char *str, size_t size)
+size_t	ft_strnlen(char *str, size_t size)
 {
 	size_t	i;
